Benthos monitor (#1965)
* savegame

* ginkgo unit tests passing

* savegame

* all unit tests passing

* all unit tests passing

* all unit tests passing

* added remaining changes

* added remaining changes

* Refactor Benthos service error handling and logging. Updated error references to use the correct service errors and improved logging levels for health check metrics. Removed unnecessary imports and ensured consistency in error messages across the Benthos service.

* Update Benthos configuration and error handling

- Increased BenthosMaxMetricsAndConfigAge from 5 seconds to 10 seconds for improved metric handling.
- Added nil check for MetricsState in HasProcessingActivity method to prevent potential nil pointer dereference.
- Enhanced error handling in ParseMetricsFromBytes and Status methods to provide clearer error messages.
- Updated mock service to ensure readiness status reflects actual connection states.

* do not try to fetch metrics when the fsm is stopped

* Update health check handling in Benthos service status

- Modify getServiceStatus method to return updated health check status when the Benthos monitor is not running and the FSM is in a starting or stopped state. This ensures that the health check accurately reflects the service's operational state.

* fix: unit test

* Refactor Benthos monitor update process in S6 manager

- Updated the UpdateBenthosMonitorInS6Manager method to generate the S6 config for the Benthos instance instead of removing and adding it.
- Enhanced error handling for S6 config generation.
- Improved the ForceRemove method in the S6 service to include log directory cleanup with error reporting.

* Add S6 manager integration to Benthos service

- Introduced WithS6Manager option to set a custom S6 manager for the Benthos service.
- Enhanced unit tests to verify coordination between Benthos and Benthos-monitor, including stopping and starting the monitor and handling configuration changes.
- Updated mock service to track the last used port for the Benthos monitor.

* added benthos_monitor fsm

* savegame

* savegame

* unti tests passing

* fix: integration test

* parallel benthos monitor processing with ctx

* feat: updated test to test benthos_monitor

* fix: port check

* more stable parsing

* fix: reset found before checking metrics

* fix: actually assign nil

* fix: set IsRunning

* fix: observe shouldnt set agent

* fix: service name

* feat: log manager time on warn

* feat: log manager time on warn

* feat: add logFilesize to S6ServiceConfig and update BenthosMonitorService configuration

- Introduced a new field `LogFilesize` in `S6ServiceConfig` to specify the maximum log file size, enhancing performance for log reading.
- Updated `GenerateS6ConfigForBenthosMonitor` to set a default log file size of 16KB, optimizing log handling for monitoring services.

* fix: permissions

* log filesize working

* smaller log file sizes for benthos monitor

* parallel base

* Revert "parallel base"

This reverts commit c7976ab22298f003e977c7cd9e5d528da5839605.

* feat: faster log line parse

* feat: faster log line parse

* feat: faster log line parse

* fix: license eye

* feat: enhance context handling and metrics tracking

- Added context cancellation checks in the GetConfig method to ensure proper error handling when the context is cancelled.
- Introduced metrics tracking for the Status method in AgentMonitorService to observe reconciliation time, improving performance monitoring capabilities.

* added niceness

* refactor: disable sys_nice capability by default in container build

- Commented out the '--cap-add=sys_nice' option in the BuildAndRunContainer function to prevent its default usage.

* removed nice entirely

* better cpu calculation

* fix: update log variable name for clarity in deploy-dataflowcomponent

* increased time for get config

* feat: bench

* fix: increase ConfigGetConfigTimeout to 20ms to prevent blocking during configuration retrieval

---------

+++ umh-core/pkg/config/benthosserviceconfig/normalizer_test.go
@@ -35,7 +35,7 @@ var _ = Describe("Benthos YAML Normalizer", func() {
 			Expect(normalizedInput).NotTo(BeNil())
 			Expect(normalizedOutput).NotTo(BeNil())
 			Expect(normalizedPipeline).NotTo(BeNil())
-			Expect(normalizedMetricsPort).To(Equal(4195))
+			Expect(normalizedMetricsPort).To(Equal(uint16(4195)))
 			Expect(normalizedLogLevel).To(Equal("INFO"))
 		})
 
@@ -60,7 +60,7 @@ var _ = Describe("Benthos YAML Normalizer", func() {
 						},
 					},
 				},
-				MetricsPort: 8000,
+				MetricsPort: uint16(8000),
 				LogLevel:    "DEBUG",
 			}
 
@@ -72,7 +72,7 @@ var _ = Describe("Benthos YAML Normalizer", func() {
 			normalizedMetricsPort := normalizedConfig.MetricsPort
 			normalizedLogLevel := normalizedConfig.LogLevel
 
-			Expect(normalizedMetricsPort).To(Equal(8000))
+			Expect(normalizedMetricsPort).To(Equal(uint16(8000)))
 			Expect(normalizedLogLevel).To(Equal("DEBUG"))
 
 			// Check input preserved
